Add interactivity to update status
