+++ src/index.html
@@ -28,7 +28,7 @@
             <img src="./assets/return.svg" alt="return icon" class="return" />
           </div>
         </form>
-        <!-- <hr /> -->
+
         <div class="list__container">
           <ul class="todo__list">
             <!-- Dynamicaly Generated -->
@@ -37,7 +37,7 @@
       </header>
 
       <div class="footer">
-        <h3 class="clearMarked">Clear all completed</h3>
+        <button class="clearMarked h3">Clear all completed</button>
       </div>
     </main>
   </body>
