+++ src/module/AddRemove.js
@@ -2,73 +2,8 @@ const todoListContianer = document.querySelector('.todo__list');
 const listConatiner = document.querySelector('.list__container');
 const singleList = document.createElement('li');
 export const clearComoleted = document.querySelector('.clearMarked');
-export const inputTodo = document.querySelector('.add__list');
 
-export default class TodoList {
-  constructor() {
-    this.list = localStorage.getItem('todoList')
-      ? JSON.parse(localStorage.getItem('todoList'))
-      : [];
-  }
-
-  addTodo() {
-    const listLength = this.list.length;
-    if (inputTodo.value !== '') {
-      this.list.push({
-        description: inputTodo.value,
-        completed: false,
-        index: listLength,
-      });
-    }
-
-    this.reArrange();
-    localStorage.setItem('todoList', JSON.stringify(this.list));
-    inputTodo.value = '';
-  }
-
-  markList(content, index, marked) {
-    if (marked === true) {
-      this.list[index].completed = true;
-      const p = content.parentNode.querySelector('.tagP');
-      p.classList.add('strike');
-    } else {
-      this.list[index].completed = false;
-      const p = content.parentNode.querySelector('.tagP');
-      p.classList.remove('strike');
-    }
-    const hold = this.list.filter((item) => item);
-    this.list = hold;
-    localStorage.setItem('todoList', JSON.stringify(this.list));
-  }
-
-  removeList() {
-    const strickers = document.querySelectorAll('.strike');
-    strickers.forEach((value) => {
-      const parentContainerLi = value.parentNode.parentNode;
-      parentContainerLi.style.display = 'none';
-      const title = parentContainerLi.querySelector('.tagP').textContent;
-      this.list = this.list.filter((value) => value.description !== title);
-    });
-
-    this.reArrange();
-    localStorage.setItem('todoList', JSON.stringify(this.list));
-  }
-
-  // SMALL UTILITIES
-
-  editListWrite(pDots, index) {
-    this.list[index].description = pDots.innerText;
-    localStorage.setItem('todoList', JSON.stringify(this.list));
-  }
-
-  reArrange() {
-    this.list.forEach((value, index) => {
-      value.index = index + 1;
-    });
-  }
-}
-
-export function render(member) {
+export function render(member, TodoListObj) {
   todoListContianer.innerHTML = '';
   member.forEach((list) => {
     const check = list.completed ? 'checked' : null;
@@ -92,6 +27,19 @@ export function render(member) {
   });
 
   // EventListner for CheckBox
+  const checkBox = document.querySelectorAll('.chBox');
+  let marked = false;
+  checkBox.forEach((content, index) => {
+    content.addEventListener('change', () => {
+      if (content.checked) {
+        marked = true;
+        TodoListObj.markList(content, index, marked);
+      } else {
+        marked = false;
+        TodoListObj.markList(content, index, marked);
+      }
+    });
+  });
 }
 
 export function editDescription(dotValue, index, TodoListObj) {
